refactor: Merge ContentModel with ContentTrackerModel
